Surprise-warning removed.

+++ server/inc/core/consoles/DataConsole.hpp
@@ -5,7 +5,7 @@
 // Login   <aracthor@epitech.net>
 // 
 // Started on  Wed Nov 19 09:14:18 2014 
-// Last Update Thu Nov 20 12:27:33 2014 
+// Last Update Tue Oct 27 14:42:09 2015
 //
 
 template <typename T>
@@ -24,7 +24,7 @@ template <typename T>
 void
 DataConsole<T>::drawElem(const T& elem, unsigned int index) const
 {
-  if (index + 1 < this->getHeight() - 1)
+  if (static_cast<int>(index) + 1 < this->getHeight() - 1)
     {
       this->move(1, index + 1);
       this->drawElemData(elem);
